+++ presets/array-nonindex-keys.js
@@ -1,0 +1,59 @@
+export default [
+    {
+        arrayNonindexKeys: {
+            testPlainObjects: true,
+            test (x, stateObj) {
+                if (Array.isArray(x)) {
+                    if (
+                        // By avoiding serializing arrays into objects which
+                        //  have only positive-integer keys, we reduce
+                        //  size and improve revival performance; arrays with
+                        //  non-index keys will be larger however
+                        Object.keys(x).some((k) => {
+                            //  No need to check for `isNaN` or
+                            //   `isNaN(parseInt())` as `NaN` will be treated as a
+                            //   string.
+                            //  No need to do check as `parseInt(Number())` since
+                            //   scientific notation will be pre-resolved if a
+                            //   number was given, and it will otherwise be a string
+                            return String(parseInt(k)) !== k;
+                        })
+                    ) {
+                        stateObj.iterateIn = 'object';
+                        stateObj.addLength = true;
+                    }
+                    return true;
+                }
+                return false;
+            },
+            replace (a, stateObj) {
+                if (Array.isArray(a)) {
+                    // Catch sparse undefined
+                    stateObj.iterateUnsetNumeric = true;
+                }
+                return a;
+            },
+            revive (o) {
+                if (Array.isArray(o)) {
+                    return o;
+                }
+                const arr = [];
+                // No map here as may be a sparse array (including
+                //   with `length` set)
+                Object.entries(o).forEach(([key, val]) => {
+                    arr[key] = val;
+                });
+                return arr;
+            }
+        }
+    },
+    {
+        sparseUndefined: {
+            test (x, stateObj) {
+                return typeof x === 'undefined' && stateObj.ownKeys === false;
+            },
+            replace (n) { return 0; },
+            revive (s) { return undefined; } // Will avoid adding anything
+        }
+    }
+];