+++ presets/sparse-undefined.js
@@ -14,7 +14,7 @@ export default [
             test (x, stateObj) {
                 return typeof x === 'undefined' && stateObj.ownKeys === false;
             },
-            replace (n) { return null; },
+            replace (n) { return 0; },
             revive (s) { return undefined; } // Will avoid adding anything
         }
     }